requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pytesseract==0.3.13
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.6
